Add BPMN DI layout generation at the end of the Simod run. Closes #154
The current problem was that (a) SplitMiner CLI tool didn't generate BPMN DI part of the BPMN; (b) during the extraneous timers discovery new elements are added to the control flow, so a new layout generation has been required.

This commit adds a Java dependency to run on a BPMN file with the control flow to add the generated diagram in the XML format.